initialize Analytiks in the Application class
