+++ app/src/main/java/com/logitanalyticsapp/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.analytiks.Analytiks
 import com.analytiks.addon.mixpanel.MixpanelAnalyticsClient
-import com.analytiks.addon.timber.TimberLocalClient
 import com.analytiks.core.model.Param
 import com.logitanalyticsapp.databinding.ActivityMainBinding
 
@@ -20,23 +19,11 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
         setSupportActionBar(binding.toolbar)
 
-        val clients = listOf(
-            TimberLocalClient(),
-            MixpanelAnalyticsClient(
-                token = "test-key-goes-here",
-                optOutTrackingDefault = true,
-            ),
-        )
-
-        analytiks = Analytiks(clients).also {
-            it.initialize(this@MainActivity)
-        }
-
-        analytiks.logFirstEvent()
+        /*analytiks.logFirstEvent()
 
         binding.fab.setOnClickListener {
             analytiks.logEventOnClick()
-        }
+        }*/
     }
 
     private fun Analytiks.logFirstEvent() {
